feat: Add prefered_login_method prompt and setting
This commit adds a new prompt and setting for the preferred login method in the `add-org.js` configuration file. The prompt asks the user to specify their preferred login method, which can be one of: email, phone_number, username, or all. The default value is set to 'all'. The corresponding setting is also added to the `default.yml` and `config.yml.hbs` files.